backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
passlib[bcrypt]
python-jose
python-multipart
email-validator
pydantic
